requirements.txt
openai
langchain-core
python-dotenv
langgraph